+++ ergo-lib/src/eval.rs
@@ -1,19 +1,20 @@
 //! Interpreter
 use crate::{
-    ast::{ops::BinOp, ops::NumOp, Constant, ConstantVal, Expr},
+    ast::{Constant, ConstantVal, Expr},
     sigma_protocol::sigma_boolean::SigmaBoolean,
     types::SType,
 };
 
 use cost_accum::CostAccumulator;
 use thiserror::Error;
-use value::Value;
 
-mod cost_accum;
+use self::context::Context;
+
+pub(crate) mod context;
+pub(crate) mod cost_accum;
 mod costs;
-mod value;
 
-/// Environment vars for script interpreter
+/// Environment for the interpreter
 pub struct Env();
 
 impl Env {
@@ -35,61 +36,101 @@ pub enum EvalError {
     UnexpectedExpr,
 }
 
-/// Result of ErgoTree reduction procedure (see `reduce_to_crypto`).
+/// Result of expression reduction procedure (see `reduce_to_crypto`).
 pub struct ReductionResult {
     /// value of SigmaProp type which represents a statement verifiable via sigma protocol.
     pub sigma_prop: SigmaBoolean,
-    /// estimated cost of contract execution
+    /// estimated cost of expression evaluation
     pub cost: u64,
 }
 
 /// Interpreter
 pub trait Evaluator {
-    /// This method is used in both prover and verifier to compute SigmaBoolean value.
-    fn reduce_to_crypto(&self, expr: &Expr, env: &Env) -> Result<ReductionResult, EvalError> {
+    /// Evaluate the given expression by reducing it to SigmaBoolean value.
+    fn reduce_to_crypto(
+        &self,
+        expr: &Expr,
+        env: &Env,
+        ctx: &Context,
+    ) -> Result<ReductionResult, EvalError> {
         let mut ca = CostAccumulator::new(0, None);
-        eval(expr, env, &mut ca).and_then(|v| match v {
-            Value::Boolean(b) => Ok(ReductionResult {
-                sigma_prop: SigmaBoolean::TrivialProp(b),
-                cost: 0,
-            }),
-            Value::SigmaProp(sb) => Ok(ReductionResult {
-                sigma_prop: *sb,
-                cost: 0,
-            }),
-            _ => Err(EvalError::InvalidResultType),
+        eval(expr, env, &mut ca, ctx).and_then(|v| -> Result<ReductionResult, EvalError> {
+            match v {
+                Constant {
+                    tpe: SType::SBoolean,
+                    v: ConstantVal::Boolean(b),
+                } => Ok(ReductionResult {
+                    sigma_prop: SigmaBoolean::TrivialProp(b),
+                    cost: 0,
+                }),
+                Constant {
+                    tpe: SType::SSigmaProp,
+                    v: ConstantVal::SigmaProp(sp),
+                } => Ok(ReductionResult {
+                    sigma_prop: sp.value().clone(),
+                    cost: 0,
+                }),
+                _ => Err(EvalError::InvalidResultType),
+            }
         })
     }
 }
 
+/// Expression evaluation.
+/// Should be implemented by every node that can be evaluated.
+pub trait Evaluable {
+    /// Evaluation routine to be implement by each node
+    fn eval(
+        &self,
+        env: &Env,
+        ca: &mut CostAccumulator,
+        ctx: &Context,
+    ) -> Result<Constant, EvalError>;
+}
+
 #[allow(unconditional_recursion)]
-fn eval(expr: &Expr, env: &Env, ca: &mut CostAccumulator) -> Result<Value, EvalError> {
+fn eval(
+    expr: &Expr,
+    env: &Env,
+    ca: &mut CostAccumulator,
+    ctx: &Context,
+) -> Result<Constant, EvalError> {
     match expr {
-        Expr::Const(Constant {
-            tpe: SType::SBoolean,
-            v: ConstantVal::Boolean(b),
-        }) => Ok(Value::Boolean(*b)),
-        Expr::Const(Constant {
-            tpe: SType::SSigmaProp,
-            v: ConstantVal::SigmaProp(sp),
-        }) => Ok(Value::SigmaProp(Box::new((*sp.value()).clone()))),
+        Expr::Const(c) => Ok(c.clone()),
         Expr::Coll { .. } => todo!(),
         Expr::Tup { .. } => todo!(),
         Expr::PredefFunc(_) => todo!(),
         Expr::CollM(_) => todo!(),
         Expr::BoxM(_) => todo!(),
-        Expr::CtxM(_) => todo!(),
+        Expr::CtxM(v) => v.eval(env, ca, ctx),
         Expr::MethodCall { .. } => todo!(),
-        Expr::BinOp(bin_op, l, r) => {
-            let v_l = eval(l, env, ca)?;
-            let v_r = eval(r, env, ca)?;
+        Expr::BinOp(_bin_op, l, r) => {
+            let _v_l = eval(l, env, ca, ctx)?;
+            let _v_r = eval(r, env, ca, ctx)?;
             ca.add_cost_of(expr);
-            Ok(match bin_op {
-                BinOp::Num(op) => match op {
-                    NumOp::Add => v_l + v_r,
-                },
-            })
+            todo!()
+            // Ok(match bin_op {
+            //     BinOp::Num(op) => match op {
+            //         NumOp::Add => v_l + v_r,
+            //     },
+            // })
         }
         _ => Err(EvalError::UnexpectedExpr),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::ast::ContextMethods;
+    use crate::ast::TryExtractFrom;
+
+    use super::*;
+
+    #[test]
+    fn height() {
+        let expr = Expr::CtxM(ContextMethods::Height);
+        let mut ca = CostAccumulator::new(0, None);
+        let res = eval(&expr, &Env::empty(), &mut ca, &Context::dummy()).unwrap();
+        assert_eq!(i32::try_extract_from(res).unwrap(), 0);
     }
 }